Use 64 bit particle indices for the sightline neighbour lists
NearParticles.part was a vector<int>, so the particle indices found by
get_near_particles were silently truncated for a snapshot with more than
2^31 particles, even though the interface takes the particle count as a
long long. part_int then computed Pos[3*ipart+axis-1] in int arithmetic,
which is signed overflow above around 716 million particles and reads
from an arbitrary offset rather than failing.

Widen part to long long, and with it the index variables that are
derived from it in part_int and assign_cells. The loop over the lines
becomes long long as well, so that the offset into the output, i*nbins,
no longer wraps when the number of lines times the number of bins
exceeds 2^32.

+++ fake_spectra/index_table.h
@@ -18,7 +18,8 @@ struct NearParticles
     //One more than the number of lines: offsets[i+1] ends line i.
     std::vector<long long> offsets;
     //Particle index, and its squared distance from the line.
-    std::vector<int> part;
+    //The index is 64 bit: a snapshot may have more than 2^31 particles.
+    std::vector<long long> part;
     std::vector<double> dr2;
 
     //Number of particles near line i
